Add logic for 2nd #, 2nd operator result, equality
Change logic so displays a result if another
operator is entered after the 2nd number instead
of an equals sign.

Add boolean logic for operator handler so that
when a operator is pressed, the operator handler
processes the 2nd digit.

Add logic for the equality sign.

Add clear button eventlistener

Add error for dot handler